For messages having sources that includes user name and host name, those are now put into the :source-user and :source-host keys. :source will now only contain the nickname.
Nicknames are the only "entity identifiers" that can have optional parts (user name and host name). The optional parts must also be omitted when a nickname is used as a target. I therefore analyze the bare nicknames to be the "real" identifiers and the optional parts to be some kind of metadata.

Also, the names of the new keys start with "source-" to signal that they belong together with the source.